[AMF] Follow-up on #3380

+++ lib/core/ogs-process.c
@@ -371,13 +371,15 @@ int ogs_proc_join(ogs_proc_t *const process, int *const out_return_code)
 
     if (process->child != waitpid(process->child, &status, 0)) {
         process->child = 0;
-        ogs_fatal("waitpid failed: %d", status);
+        ogs_error("waitpid failed: %d", status);
         return OGS_ERROR;
     }
 
     process->child = 0;
-    if (process->nf_name)
+    if (process->nf_name) {
         ogs_free(process->nf_name);
+        process->nf_name = NULL;
+    }
 
     if (out_return_code) {
         if (WIFEXITED(status)) {
@@ -425,6 +427,10 @@ int ogs_proc_terminate(ogs_proc_t *const process)
 
     if (process->child) {
         if (kill(process->child, SIGTERM) == -1) {
+            if (errno == ESRCH) {
+                /* No such process */
+                return OGS_OK;
+            }
             return OGS_ERROR;
         }
     }
@@ -444,8 +450,14 @@ int ogs_proc_kill(ogs_proc_t *const process)
         return OGS_ERROR;
     }
 #else
-    if (kill(process->child, SIGKILL) == -1) {
-        return OGS_ERROR;
+    if (process->child) {
+        if (kill(process->child, SIGKILL) == -1) {
+            if (errno == ESRCH) {
+                /* No such process */
+                return OGS_OK;
+            }
+            return OGS_ERROR;
+        }
     }
 #endif
